+++ src/utils.h
@@ -19,7 +19,7 @@
 
 typedef struct {
   int r, c;  // rows and columns
-  float val; // non-zero float value
+  double val; // non-zero double value
 } record_t;
 
 bool cmp_key_row(const record_t &a, const record_t &b) { return a.r < b.r; }
@@ -45,7 +45,7 @@ void get_matrix_size(const char *file_name, int &M, int &N, int &nz) {
 }
 
 // nz: number of non-zero elems, I: x-axis, J: y-axis, val: value
-void get_matrix(const char *file_name, int *I, int *J, float *val) {
+void get_matrix(const char *file_name, int *I, int *J, double *val) {
   std::ifstream in;
   in.open(file_name);
   if (!in.is_open()) {
@@ -63,7 +63,7 @@ void get_matrix(const char *file_name, int *I, int *J, float *val) {
 
   int i = 0;
   while (std::getline(in, line) && line != "") {
-    sscanf(line.c_str(), "%d %d %f", &I[i], &J[i], &val[i]);
+    sscanf(line.c_str(), "%d %d %lf", &I[i], &J[i], &val[i]);
     I[i] -= 1;
     J[i] -= 1;
     ++i;
@@ -91,7 +91,7 @@ void get_records(const char *file_name, record_t *records) {
   int count = 0;
   record_t tmp;
   while (std::getline(in, line) && line != "") {
-    sscanf(line.c_str(), "%d %d %f", &tmp.r, &tmp.c, &tmp.val);
+    sscanf(line.c_str(), "%d %d %lf", &tmp.r, &tmp.c, &tmp.val);
     tmp.r -= 1;
     tmp.c -= 1;
     records[count] = tmp;
@@ -109,7 +109,7 @@ void records_reorder_by_columns(const int &nz, record_t *records) {
 }
 
 // source vector x random generation
-void rand_gen(const int &len, float *x) {
+void rand_gen(const int &len, double *x) {
   static thread_local std::mt19937 seed;
   std::uniform_real_distribution<> dis(-RAND_MAX, RAND_MAX);
   for (int i = 0; i < len; ++i) {
@@ -119,7 +119,7 @@ void rand_gen(const int &len, float *x) {
 }
 
 // check the correctness of optimizer output and naive result
-bool check(const size_t &len, float *output, float *result) {
+bool check(const size_t &len, double *output, double *result) {
   bool pass = true;
   for (std::size_t i = 0; i < len; ++i) {
     if ((output[i] - result[i]) > ESP) {
@@ -130,7 +130,7 @@ bool check(const size_t &len, float *output, float *result) {
   return pass;
 }
 
-void cvt2csr(const int &rows, const int &nz, record_t *records, float *nz_vals,
+void cvt2csr(const int &rows, const int &nz, record_t *records, double *nz_vals,
              int *column_index, int *row_start) {
   // TODO: fill with a bit map
   int *row_tmp = (int *)malloc(rows * sizeof(int));
@@ -168,7 +168,7 @@ void cvt2csr(const int &rows, const int &nz, record_t *records, float *nz_vals,
 }
 
 void cvt2csc(const int &columns, const int &nz, record_t *records,
-             float *nz_vals, int *row_index, int *column_start) {
+             double *nz_vals, int *row_index, int *column_start) {
   // TODO: fill with a bit map
   int *col_tmp = (int *)malloc(columns * sizeof(int));
   if (!col_tmp) {
@@ -209,10 +209,10 @@ void cvt2csc(const int &columns, const int &nz, record_t *records,
 
 #if 0
 // convert the input matrix into csr format
-void cvt2csr(const int &rows, const int &columns, const int &nz, float *A,
-             float *nz_vals, int *column_index, int *row_start) {
+void cvt2csr(const int &rows, const int &columns, const int &nz, double *A,
+             double *nz_vals, int *column_index, int *row_start) {
   int count = 0;
-  float element;
+  double element;
   for (int i = 0; i < rows; ++i) {
     row_start[i] = count;
     for (int j = 0; j < columns; ++j) {
@@ -228,10 +228,10 @@ void cvt2csr(const int &rows, const int &columns, const int &nz, float *A,
 }
 
 // convert the input matrix into csc format
-void cvt2csc(const int &rows, const int &columns, const int &nz, float *A,
-             float *nz_vals, int *row_index, int *column_start) {
+void cvt2csc(const int &rows, const int &columns, const int &nz, double *A,
+             double *nz_vals, int *row_index, int *column_start) {
   int count = 0;
-  float element;
+  double element;
   for (int j = 0; j < columns; ++j) {
     column_start[j] = count;
     for (int i = 0; i < rows; ++i) {
